+++ app/designs/[id]/layout.tsx
@@ -1,0 +1,11 @@
+export default function ProjectLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="min-h-screen bg-background text-foreground m-10">
+      {children}
+    </div>
+  );
+}
