+++ components/designs-gallery.tsx
@@ -170,7 +170,6 @@ export default function DesignsGallery() {
         >
           <Link
             href={`/designs/${design.id}`}
-            target="_blank"
             rel="noopener noreferrer"
           >
             <div className="relative w-full aspect-video rounded-lg border overflow-hidden">
