+++ package.json
@@ -46,6 +46,7 @@
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-draggable": "^4.5.0",
     "react-hook-form": "^7.60.0",
     "recharts": "^2.15.4",
     "sonner": "^2.0.7",
